Update Proguard rules for data classes
This commit updates the Proguard rules for classes in the `com.kouloundissa.twinstracker.data` package.
The rule is expanded to keep all methods (`<methods>`) and everything else (`*`) within these classes, in addition to constructors and fields.

This change ensures that all necessary members of the data classes are preserved during obfuscation and shrinking, which is crucial for libraries like Kotlin serialization to function correctly.